pyproject.toml
[tool.poetry]
name = "officemanagement"
version = "0.1.0"
description = "Scientific office management web platform"
authors = ["Mykola Subota <subotamykola121@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
djangorestframework = "^3.12.2"
Pillow = "^8.0.1"
psycopg2-binary = "^2.8.6"
djoser = "^2.1.0"
djangorestframework-simplejwt = "^4.6.0"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.6.0"
django-mptt = "^0.11.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
